style: separate stats code into more functions
