+++ pages/api/stats.ts
@@ -1,34 +1,10 @@
 import { withApiMethods } from "../../lib/withApiMethods";
-import { prisma } from "../../lib/db";
-import { generateStats } from "../../lib/stats/stats";
+import * as stats from "../../lib/stats/stats";
 
 export default withApiMethods({
   GET: async (req, res) => {
-    const stats = await prisma.stat.findMany();
+    const newStats = stats.update();
 
-    if (!stats.length || stats[0]?.updatedAt < new Date(Date.now() - 1000 * 10)) {
-      const newStats = await generateStats();
-
-      const upserts = Object.entries(newStats).map(([key, value]) =>
-        prisma.stat.upsert({
-          where: {
-            id: key,
-          },
-          update: {
-            value: String(value),
-            updatedAt: new Date(Date.now()),
-          },
-          create: {
-            id: key,
-            value: String(value),
-            updatedAt: new Date(Date.now()),
-          },
-        }),
-      );
-
-      await prisma.$transaction(upserts);
-    }
-
-    return res.json(Object.fromEntries(stats.map((stat) => [stat.id, stat.value])));
+    return res.json(newStats);
   },
 });
